- drop old jit executor
