+++ src/core/app_context.rs
@@ -1,12 +1,10 @@
 use std::sync::Arc;
 
 use async_graphql::dynamic::{self, DynamicRequest};
-use async_graphql::Response;
 use async_graphql_value::ConstValue;
 use dashmap::DashMap;
 
 use super::jit::ByteResponse;
-use super::lift::Lift;
 use crate::core::async_graphql_hyper::OperationId;
 use crate::core::auth::context::GlobalAuthContext;
 use crate::core::blueprint::{Blueprint, Definition, SchemaModifiers};
@@ -31,8 +29,7 @@ pub struct AppContext {
     pub endpoints: EndpointSet<Checked>,
     pub auth_ctx: Arc<GlobalAuthContext>,
     pub dedupe_handler: Arc<DedupeResult<IoId, ConstValue, Error>>,
-    pub dedupe_operation_handler: DedupeResult<OperationId, Lift<Response>, Error>,
-    pub dedupe_operation_handler_arc: DedupeResult<OperationId, Arc<ByteResponse>, Error>,
+    pub dedupe_operation_handler: DedupeResult<OperationId, Arc<ByteResponse>, Error>,
     pub operation_plans: DashMap<OPHash, OperationPlan<async_graphql_value::Value>>,
 }
 
@@ -154,7 +151,6 @@ impl AppContext {
             auth_ctx: Arc::new(auth_ctx),
             dedupe_handler: Arc::new(DedupeResult::new(false)),
             dedupe_operation_handler: DedupeResult::new(false),
-            dedupe_operation_handler_arc: DedupeResult::new(false),
             operation_plans: DashMap::new(),
         }
     }
