+++ src/core/jit/graphql_executor.rs
@@ -3,16 +3,17 @@ use std::future::Future;
 use std::hash::{Hash, Hasher};
 use std::sync::Arc;
 
-use async_graphql::{Data, Executor, Response, Value};
+use async_graphql::{BatchRequest, Response, Value};
 use async_graphql_value::{ConstValue, Extensions};
-use futures_util::stream::BoxStream;
+use futures_util::stream::FuturesOrdered;
+use futures_util::StreamExt;
 use tailcall_hasher::TailcallHasher;
 
+use super::{BatchResponse, ByteResponse};
 use crate::core::app_context::AppContext;
 use crate::core::async_graphql_hyper::OperationId;
 use crate::core::http::RequestContext;
 use crate::core::jit::{self, ConstValueExecutor, OPHash};
-use crate::core::lift::{CanLift, Lift};
 use crate::core::merge_right::MergeRight;
 
 #[derive(Clone)]
@@ -36,20 +37,22 @@ impl JITExecutor {
         &self,
         exec: ConstValueExecutor,
         jit_request: jit::Request<ConstValue>,
-    ) -> Response {
+    ) -> Arc<ByteResponse> {
         let is_introspection_query = self.app_ctx.blueprint.server.get_enable_introspection()
             && exec.plan.is_introspection_query;
         let jit_resp = exec
             .execute(&self.req_ctx, &jit_request)
             .await
             .into_async_graphql();
-        if is_introspection_query {
+        let response = if is_introspection_query {
             let async_req = async_graphql::Request::from(jit_request).only_introspection();
             let async_resp = self.app_ctx.execute(async_req).await;
             jit_resp.merge_right(async_resp)
         } else {
             jit_resp
-        }
+        };
+
+        Arc::new(response.into())
     }
 
     #[inline(always)]
@@ -57,19 +60,19 @@ impl JITExecutor {
         &self,
         exec: ConstValueExecutor,
         jit_request: jit::Request<ConstValue>,
-    ) -> Response {
+    ) -> Arc<ByteResponse> {
         let out = self
             .app_ctx
             .dedupe_operation_handler
             .dedupe(&self.operation_id, || {
                 Box::pin(async move {
                     let resp = self.exec(exec, jit_request).await;
-                    Ok(resp.lift())
+                    Ok(resp)
                 })
             })
             .await;
 
-        out.map(|response| response.take()).unwrap_or_default()
+        out.unwrap_or_default()
     }
 
     #[inline(always)]
@@ -81,14 +84,50 @@ impl JITExecutor {
     }
 }
 
-impl Clone for Lift<async_graphql::Response> {
-    fn clone(&self) -> Self {
-        let mut res = async_graphql::Response::new(self.data.clone())
-            .cache_control(self.cache_control)
-            .http_headers(self.http_headers.clone());
-        res.errors = self.errors.clone();
-        res.extensions = self.extensions.clone();
-        res.into()
+impl JITExecutor {
+    pub fn execute(
+        &self,
+        request: async_graphql::Request,
+    ) -> impl Future<Output = Arc<ByteResponse>> + Send + '_ {
+        let hash = Self::req_hash(&request);
+
+        async move {
+            let jit_request = jit::Request::from(request);
+            let mut exec = if let Some(op) = self.app_ctx.operation_plans.get(&hash) {
+                ConstValueExecutor::from(op.value().clone())
+            } else {
+                let exec = match ConstValueExecutor::try_new(&jit_request, &self.app_ctx) {
+                    Ok(exec) => exec,
+                    Err(error) => {
+                        return Arc::new(Response::from_errors(vec![error.into()]).into())
+                    }
+                };
+                self.app_ctx.operation_plans.insert(hash, exec.plan.clone());
+                exec
+            };
+
+            if let Some(response) = std::mem::take(&mut exec.response) {
+                Arc::new(response.into_async_graphql().into())
+            } else if exec.plan.is_query() && exec.plan.dedupe {
+                self.dedupe_and_exec(exec, jit_request).await
+            } else {
+                self.exec(exec, jit_request).await
+            }
+        }
+    }
+
+    /// Execute a GraphQL batch query.
+    pub async fn execute_batch(&self, batch_request: BatchRequest) -> BatchResponse {
+        match batch_request {
+            BatchRequest::Single(request) => BatchResponse::Single(self.execute(request).await),
+            BatchRequest::Batch(requests) => {
+                let futs = FuturesOrdered::from_iter(
+                    requests.into_iter().map(|request| self.execute(request)),
+                );
+                let responses = futs.collect::<Vec<_>>().await;
+                BatchResponse::Batch(responses)
+            }
+        }
     }
 }
 
@@ -106,41 +145,5 @@ impl From<jit::Request<Value>> for async_graphql::Request {
         request.extensions = Extensions(value.extensions);
         request.operation_name = value.operation_name;
         request
-    }
-}
-
-impl Executor for JITExecutor {
-    fn execute(&self, request: async_graphql::Request) -> impl Future<Output = Response> + Send {
-        let hash = Self::req_hash(&request);
-
-        async move {
-            let jit_request = jit::Request::from(request);
-            let mut exec = if let Some(op) = self.app_ctx.operation_plans.get(&hash) {
-                ConstValueExecutor::from(op.value().clone())
-            } else {
-                let exec = match ConstValueExecutor::try_new(&jit_request, &self.app_ctx) {
-                    Ok(exec) => exec,
-                    Err(error) => return Response::from_errors(vec![error.into()]),
-                };
-                self.app_ctx.operation_plans.insert(hash, exec.plan.clone());
-                exec
-            };
-
-            if let Some(response) = std::mem::take(&mut exec.response) {
-                response.into_async_graphql()
-            } else if exec.plan.is_query() && exec.plan.dedupe {
-                self.dedupe_and_exec(exec, jit_request).await
-            } else {
-                self.exec(exec, jit_request).await
-            }
-        }
-    }
-
-    fn execute_stream(
-        &self,
-        _: async_graphql::Request,
-        _: Option<Arc<Data>>,
-    ) -> BoxStream<'static, Response> {
-        unimplemented!("streaming not supported")
     }
 }
