+++ src/core/jit/mod.rs
@@ -16,13 +16,11 @@ mod response;
 mod builder;
 pub mod common;
 mod graphql_executor;
-mod graphql_executor_arc;
 
 // Public Exports
 pub use error::*;
 pub use exec_const::*;
 pub use graphql_executor::*;
-pub use graphql_executor_arc::*;
 pub use model::*;
 pub use request::*;
 pub use response::*;
